intial open source commit
moved much of the branded text out into a config/branding folder that is not included here and only added the default example environment config - all the rest is what is installed to chatterscan.com. This is mainly for reference, not expecting anyone to fork, use or contribute to this.